Move FieldContentParser to bibtex package
