+++ src/main/java/net/sf/jabref/bibtex/LatexFieldFormatter.java
@@ -20,7 +20,6 @@ import java.util.List;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.importer.fileformat.FieldContentParser;
 import net.sf.jabref.logic.util.strings.StringUtil;
 
 /**
